pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sesame-exporter"
version = "0.1.0"
description = "A Prometheus exporter for the Sesame API."
readme = "README.md"
requires-python = ">=3.7"
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "prometheus-client",
    "PyYAML",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
sesame-exporter = "sesame_exporter:main"

[tool.black]
line-length = 88
target-version = ['py37']

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
